catch up on sklearn testing
