+++ gplearn/genetic.py
@@ -22,6 +22,7 @@ from sklearn.base import RegressorMixin, TransformerMixin, ClassifierMixin
 from sklearn.exceptions import NotFittedError
 from sklearn.utils import compute_sample_weight
 from sklearn.utils.validation import check_X_y, check_array
+from sklearn.utils.validation import  _check_sample_weight
 from sklearn.utils.multiclass import check_classification_targets
 
 from ._program import _Program
@@ -286,6 +287,7 @@ class BaseSymbolic(BaseEstimator, metaclass=ABCMeta):
         # Check arrays
         if sample_weight is not None:
             sample_weight = check_array(sample_weight, ensure_2d=False)
+            sample_weight = _check_sample_weight(sample_weight, X)
 
         if isinstance(self, ClassifierMixin):
             X, y = check_X_y(X, y, y_numeric=False)
@@ -310,7 +312,7 @@ class BaseSymbolic(BaseEstimator, metaclass=ABCMeta):
         else:
             X, y = check_X_y(X, y, y_numeric=True)
 
-        _, self.n_features_ = X.shape
+        _, self.n_features_in_ = X.shape
 
         hall_of_fame = self.hall_of_fame
         if hall_of_fame is None:
@@ -394,10 +396,11 @@ class BaseSymbolic(BaseEstimator, metaclass=ABCMeta):
                              'order: (min_depth, max_depth).')
 
         if self.feature_names is not None:
-            if self.n_features_ != len(self.feature_names):
+            if self.n_features_in_ != len(self.feature_names):
                 raise ValueError('The supplied `feature_names` has different '
                                  'length to n_features. Expected %d, got %d.'
-                                 % (self.n_features_, len(self.feature_names)))
+                                 % (self.n_features_in_,
+                                    len(self.feature_names)))
             for feature_name in self.feature_names:
                 if not isinstance(feature_name, str):
                     raise ValueError('invalid type %s found in '
@@ -859,11 +862,11 @@ class SymbolicRegressor(BaseSymbolic, RegressorMixin):
 
         X = check_array(X)
         _, n_features = X.shape
-        if self.n_features_ != n_features:
+        if self.n_features_in_ != n_features:
             raise ValueError('Number of features of the model must match the '
                              'input. Model n_features is %s and input '
                              'n_features is %s.'
-                             % (self.n_features_, n_features))
+                             % (self.n_features_in_, n_features))
 
         y = self._program.execute(X)
 
@@ -1154,11 +1157,11 @@ class SymbolicClassifier(BaseSymbolic, ClassifierMixin):
 
         X = check_array(X)
         _, n_features = X.shape
-        if self.n_features_ != n_features:
+        if self.n_features_in_ != n_features:
             raise ValueError('Number of features of the model must match the '
                              'input. Model n_features is %s and input '
                              'n_features is %s.'
-                             % (self.n_features_, n_features))
+                             % (self.n_features_in_, n_features))
 
         scores = self._program.execute(X)
         proba = self._transformer(scores)
@@ -1454,6 +1457,15 @@ class SymbolicTransformer(BaseSymbolic, TransformerMixin):
         output = str([gp.__str__() for gp in self])
         return output.replace("',", ",\n").replace("'", "")
 
+    def _more_tags(self):
+        return {
+            "_xfail_checks": {
+                "check_sample_weights_invariance": (
+                    "zero sample_weight is not equivalent to removing samples"
+                ),
+            }
+        }
+
     def transform(self, X):
         """Transform X according to the fitted transformer.
 
@@ -1474,11 +1486,11 @@ class SymbolicTransformer(BaseSymbolic, TransformerMixin):
 
         X = check_array(X)
         _, n_features = X.shape
-        if self.n_features_ != n_features:
+        if self.n_features_in_ != n_features:
             raise ValueError('Number of features of the model must match the '
                              'input. Model n_features is %s and input '
                              'n_features is %s.'
-                             % (self.n_features_, n_features))
+                             % (self.n_features_in_, n_features))
 
         X_new = np.array([gp.execute(X) for gp in self._best_programs]).T
